refactor base class for jupyter request handling
